Link: report the heartbeat arm flag and split message folding out
The FC's heartbeat carries its arm state; the cache now retains it so
adapters can publish an FC-state sample without a second polling
socket. Message folding moves to link/apply.rs unchanged.
