+++ crates/pilotage-mavlink/src/link.rs
@@ -12,16 +12,16 @@ use tracing::{debug, error, info, warn};
 
 use pilotage_adapter_api::{MeasurementStamp, SourceIncarnation};
 
-use crate::codec::{FcMessage, FrameSource, encode_gcs_heartbeat, parse_datagram};
+use crate::codec::{encode_gcs_heartbeat, parse_datagram};
 
+mod apply;
 mod error;
 pub mod estimator;
 pub mod measurement;
+use apply::apply_messages;
+pub use apply::apply_messages_at;
 pub use error::LinkError;
-use estimator::{
-    EstimatorStatusUpdate, accept_status, authorization_at, invalidate_cached_authorization,
-};
-use measurement::{next_attitude_stamp, next_kinematics_stamp};
+use estimator::EstimatorStatusUpdate;
 
 /// Which message carries the estimator authorization for cached numeric
 /// groups.
@@ -152,6 +152,8 @@ pub struct LinkState {
     pub estimator_status: Option<EstimatorStatusUpdate>,
     /// Receive stamp of the last FC heartbeat.
     pub last_heartbeat: Option<Instant>,
+    /// Whether the last FC heartbeat reported the vehicle armed.
+    pub heartbeat_armed: Option<bool>,
     /// Total decoded frames.
     pub decoded: u64,
     /// Total CRC failures (a correctness signal, logged).
@@ -196,6 +198,7 @@ impl Default for LinkState {
             kinematics: None,
             estimator_status: None,
             last_heartbeat: None,
+            heartbeat_armed: None,
             decoded: 0,
             crc_failures: 0,
             unknown_ids: 0,
@@ -385,101 +388,6 @@ async fn run_link(
                         warn!(%error, "MAVLink socket receive failed");
                         tokio::time::sleep(Duration::from_millis(100)).await;
                     }
-                }
-            }
-        }
-    }
-}
-
-/// Folds decoded messages into the shared cache. Kept synchronous and
-/// lock-scoped: the lock is never held across an await.
-fn apply_messages(
-    state: &Arc<Mutex<LinkState>>,
-    messages: &[(FrameSource, FcMessage)],
-    crc_failures: u32,
-    unknown_ids: u32,
-) {
-    apply_messages_at(state, messages, crc_failures, unknown_ids, Instant::now());
-}
-
-/// Folds decoded messages into the shared cache at an explicit receive
-/// instant. Public so adapter crates can drive the cache in tests
-/// without a socket; production traffic arrives via the link task.
-pub fn apply_messages_at(
-    state: &Arc<Mutex<LinkState>>,
-    messages: &[(FrameSource, FcMessage)],
-    crc_failures: u32,
-    unknown_ids: u32,
-    now: Instant,
-) {
-    let Ok(mut latest) = state.lock() else {
-        return;
-    };
-    latest.crc_failures = latest.crc_failures.wrapping_add(u64::from(crc_failures));
-    latest.unknown_ids = latest.unknown_ids.wrapping_add(u64::from(unknown_ids));
-    for &(source, message) in messages {
-        if source.system_id != latest.system_id || source.component_id != latest.component_id {
-            latest.wrong_sources = latest.wrong_sources.wrapping_add(1);
-            continue;
-        }
-        if message == FcMessage::InvalidAviateEstimatorStatus {
-            latest.invalid_estimator_statuses = latest.invalid_estimator_statuses.wrapping_add(1);
-            invalidate_cached_authorization(&mut latest);
-            continue;
-        }
-        latest.decoded = latest.decoded.wrapping_add(1);
-        match message {
-            FcMessage::InvalidAviateEstimatorStatus => {}
-            FcMessage::Heartbeat { .. } => latest.last_heartbeat = Some(now),
-            FcMessage::CommandAck { .. } => {}
-            FcMessage::EstimatorStatus { time_usec, flags } => {
-                // Standard-status dialects authorize from msg 230; the
-                // Aviate dialect treats it as diagnostic only.
-                if latest.authorization_source == AuthorizationSource::StandardEstimatorStatus {
-                    let (valid_flags, quality) = estimator::standard_authorization(flags);
-                    let aligned_usec = (time_usec / 1_000) * 1_000;
-                    accept_status(&mut latest, aligned_usec, valid_flags, quality, now);
-                }
-            }
-            FcMessage::AviateEstimatorStatus {
-                time_usec,
-                valid_flags,
-                quality,
-            } => accept_status(&mut latest, time_usec, valid_flags, quality, now),
-            FcMessage::AttitudeQuaternion {
-                time_boot_ms,
-                quat_wxyz,
-                rates_rps,
-            } => {
-                if let Some(stamp) = next_attitude_stamp(&mut latest, time_boot_ms, now) {
-                    let authorization = authorization_at(&latest, time_boot_ms);
-                    latest.attitude = Some(AttitudeUpdate {
-                        quat_wxyz,
-                        rates_rps,
-                        time_boot_ms,
-                        stamp,
-                        valid_flags: authorization.valid_flags,
-                        quality: authorization.quality,
-                        received_at: now,
-                    });
-                }
-            }
-            FcMessage::LocalPositionNed {
-                time_boot_ms,
-                pos_ned_m,
-                vel_ned_mps,
-            } => {
-                if let Some(stamp) = next_kinematics_stamp(&mut latest, time_boot_ms, now) {
-                    let authorization = authorization_at(&latest, time_boot_ms);
-                    latest.kinematics = Some(KinematicsUpdate {
-                        pos_ned_m,
-                        vel_ned_mps,
-                        time_boot_ms,
-                        stamp,
-                        valid_flags: authorization.valid_flags,
-                        quality: authorization.quality,
-                        received_at: now,
-                    });
                 }
             }
         }
